+++ lib/linear_percent_indicator.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: unused_element
+
 import 'package:flutter/material.dart';
 
 @Deprecated('This property is no longer used, please use barRadius instead.')
@@ -64,6 +66,9 @@ class LinearPercentIndicator extends StatefulWidget {
   /// set true if you want to animate the linear from the last percent value you set
   final bool animateFromLastPercent;
 
+  /// set to false if you do not want the default behavior of initially animating up from 0%
+  final bool animateToInitialPercent;
+
   /// If present, this will make the progress bar colored by this gradient.
   ///
   /// This will override [progressColor]. It is an error to provide both.
@@ -95,6 +100,9 @@ class LinearPercentIndicator extends StatefulWidget {
   /// Display a widget indicator at the end of the progress. It only works when `animation` is true
   final Widget? widgetIndicator;
 
+  /// Return current percent value if animation is true.
+  final Function(double value)? onPercentValue;
+
   LinearPercentIndicator({
     Key? key,
     this.fillColor = Colors.transparent,
@@ -108,6 +116,7 @@ class LinearPercentIndicator extends StatefulWidget {
     this.animation = false,
     this.animationDuration = 500,
     this.animateFromLastPercent = false,
+    this.animateToInitialPercent = true,
     this.isRTL = false,
     this.leading,
     this.trailing,
@@ -124,6 +133,7 @@ class LinearPercentIndicator extends StatefulWidget {
     this.onAnimationEnd,
     this.widgetIndicator,
     this.progressBorderColor,
+    this.onPercentValue,
   }) : super(key: key) {
     if (linearGradient != null && progressColor != null) {
       throw ArgumentError(
@@ -181,14 +191,16 @@ class _LinearPercentIndicatorState extends State<LinearPercentIndicator>
       }
     });
     if (widget.animation) {
+      if (!widget.animateToInitialPercent) _percent = widget.percent;
       _animationController = AnimationController(
           vsync: this,
           duration: Duration(milliseconds: widget.animationDuration));
-      _animation = Tween(begin: 0.0, end: widget.percent).animate(
+      _animation = Tween(begin: _percent, end: widget.percent).animate(
         CurvedAnimation(parent: _animationController!, curve: widget.curve),
       )..addListener(() {
           setState(() {
             _percent = _animation!.value;
+            widget.onPercentValue?.call(_percent);
           });
           if (widget.restartAnimation && _percent == 1.0) {
             _animationController!.repeat(min: 0, max: 1.0);
@@ -252,51 +264,55 @@ class _LinearPercentIndicatorState extends State<LinearPercentIndicator>
     final hasSetWidth = widget.width != null;
     final percentPositionedHorizontal =
         _containerWidth * _percent - _indicatorWidth / 3;
-    var containerWidget = Container(
-      width: hasSetWidth ? widget.width : double.infinity,
-      height: widget.lineHeight,
-      padding: widget.padding,
-      child: Stack(
-        clipBehavior: Clip.none,
-        children: [
-          CustomPaint(
-            key: _containerKey,
-            painter: _LinearPainter(
-              isRTL: widget.isRTL,
-              progress: _percent,
-              progressColor: widget.progressColor,
-              progressBorderColor: widget.progressBorderColor,
-              linearGradient: widget.linearGradient,
-              backgroundColor: widget.backgroundColor,
-              barRadius: widget.barRadius ??
-                  Radius.zero, // If radius is not defined, set it to zero
-              linearGradientBackgroundColor:
-                  widget.linearGradientBackgroundColor,
-              maskFilter: widget.maskFilter,
-              clipLinearGradient: widget.clipLinearGradient,
+    //LayoutBuilder is used to get the size of the container where the widget is rendered
+    var containerWidget = LayoutBuilder(builder: (context, constraints) {
+      _containerWidth = constraints.maxWidth;
+      _containerHeight = constraints.maxHeight;
+      return Container(
+        width: hasSetWidth ? widget.width : double.infinity,
+        height: widget.lineHeight,
+        padding: widget.padding,
+        child: Stack(
+          clipBehavior: Clip.none,
+          children: [
+            CustomPaint(
+              key: _containerKey,
+              painter: _LinearPainter(
+                isRTL: widget.isRTL,
+                progress: _percent,
+                progressColor: widget.progressColor,
+                linearGradient: widget.linearGradient,
+                backgroundColor: widget.backgroundColor,
+                barRadius: widget.barRadius ??
+                    Radius.zero, // If radius is not defined, set it to zero
+                linearGradientBackgroundColor:
+                    widget.linearGradientBackgroundColor,
+                maskFilter: widget.maskFilter,
+                clipLinearGradient: widget.clipLinearGradient,
+              ),
+              child: (widget.center != null)
+                  ? Center(child: widget.center)
+                  : Container(),
             ),
-            child: (widget.center != null)
-                ? Center(child: widget.center)
-                : Container(),
-          ),
-          if (widget.widgetIndicator != null && _indicatorWidth == 0)
-            Opacity(
-              opacity: 0.0,
-              key: _keyIndicator,
-              child: widget.widgetIndicator,
-            ),
-          if (widget.widgetIndicator != null &&
-              _containerWidth > 0 &&
-              _indicatorWidth > 0)
-            Positioned(
-              right: widget.isRTL ? percentPositionedHorizontal : null,
-              left: !widget.isRTL ? percentPositionedHorizontal : null,
-              top: _containerHeight / 2 - _indicatorHeight,
-              child: widget.widgetIndicator!,
-            ),
-        ],
-      ),
-    );
+            if (widget.widgetIndicator != null && _indicatorWidth == 0)
+              Opacity(
+                opacity: 0.0,
+                key: _keyIndicator,
+                child: widget.widgetIndicator,
+              ),
+            if (widget.widgetIndicator != null &&
+                _containerWidth > 0 &&
+                _indicatorWidth > 0)
+              Positioned(
+                right: widget.isRTL ? percentPositionedHorizontal : null,
+                left: !widget.isRTL ? percentPositionedHorizontal : null,
+                top: _containerHeight / 2 - _indicatorHeight,
+                child: widget.widgetIndicator!,
+              ),
+          ],
+        ),
+      );
+    });
 
     if (hasSetWidth) {
       items.add(containerWidget);